fix: use IP-based rate limiting instead in Authorization handlers
Authorization handlers does not have account id extractor, so applying
rate limiting based on it was causing panics. This commit fixes the
issue by applying IP-based rate limiting.

feat: serialize ID's as string

64-bit integers are not supported by JavaScript JSON.parse. Serializing
ID's as strings facilities future implementations.

+++ tests/src/account_service.rs
@@ -119,7 +119,7 @@ pub async fn account_creation(account_service: Arc<AccountService>) -> ServiceRe
 
     let me = account_service.me(account_id).await?;
 
-    assert!(me.id == i64::from(account_id));
+    assert!(me.id == account_id);
     assert!(me.username.unwrap() == username);
     assert!(me.email.unwrap() == email);
     assert!(me.secondary_emails.is_empty());
